additem

+++ EVTovar/EVTovar/Models/Item.cs
@@ -8,8 +8,8 @@ namespace EVTovar.Models
     public class Item : BaseItem
     {
         public DateTime Modified { get; set; }
-        public float Weight { get; set; }
-        public int Price { get; set; }
+        public int Weight { get; set; }
+        public decimal Price { get; set; }
 
     }
 }
